semi-final
1. solve run time error with cancel self message
2. merge sender and receriver
3. take values from .ini
4. finish coordinator
5. solve the probability of loss problem

+++ src/coordinator.h
@@ -26,7 +26,7 @@ using namespace omnetpp;
 class Coordinator : public cSimpleModule
 {
   bool node_id;
-  int starting_time;
+  float starting_time;
 
 protected:
   virtual void initialize() override;
